-implemented the logic of IE screen using nested recyclerviews -next step is to add a third party library which expands a floating action button giving us the possibility to create an income or a category. -also we need to implement the delete option for the recycler view elements

+++ app/src/main/java/com/example/robi/investorsapp/database/ie/IEObject.java
@@ -26,7 +26,7 @@ public class IEObject {
     //public int icon[];
     public int type;//0 means income, 1 means expense
 
-    public IEObject(int wallet_id, double amount, String name, String category, int type) {
+    public IEObject(long wallet_id, double amount, String name, String category, int type) {
         this.id = System.nanoTime();
         this.name = name;
         this.amount = amount;
